Create README.md
This is the first iteration of an Greek *and* English combination dictionary for hunspell. There is still a lot to be desired but it works enough to be installed and used.

This repository also contains a small python script that makes this kind of merging possible for other languages as well. It is hacky and should be carefully adjusted for each specific case.